Support SwiftUI refreshable with async refresh methods
Summary:
SwiftUI in iOS 15 supports pull-to-refresh with a new `refreshable` view modifier. It takes an async closure, so whenever the async work is complete, that is when the spinner is dismissed.

To make it easy/possible to use this with Relay, we need to expose async methods for refetching data as an alternative to the existing methods.

Ideally, when using this, the existing data would remain in place and the UI wouldn't change until the new data was fetched.

Test Plan: Use RelayStarWars as a test for this by changing its existing refresh button to use `refreshable`.

Reviewers: matt

Differential Revision: https://code.home.mattmoriarity.com/D14