Actually replace edges for connection fields when refetching
Summary: Found this bug while working on {D14}. I had never actually tested refetching for a pagination fragment. The expected behavior is for a refetch to replace the previous list of edges with new edges, and it looked like the code did this. But the checks for whether a before or after cursor was used did not correctly check for null values, so they were hitting the wrong case and essentially doing nothing.

Test Plan: Added a unit test for this case, and also confirmed on the branch for D14 that this works as expected.

Reviewers: matt

Reviewed By: matt

Differential Revision: https://code.home.mattmoriarity.com/D15