Add Store parameter for GC release buffer size
Summary: Adds passthrough parameter to Store to configure the size of the GC release buffer. The GC has supported keeping a buffer of operations to continue retaining records for even after they are no longer retained, but there was no way to configure it to be anything besides the default of 0.

Test Plan: The example apps are too small for testing this, so I may need some help exercising this to see if it works as expected.

Reviewers: matt

Differential Revision: https://code.home.mattmoriarity.com/D16