Support SwiftUI refreshable with async refresh methods
Summary:
SwiftUI in iOS 15 supports pull-to-refresh with a new `refreshable` view modifier. It takes an async closure, so whenever the async work is complete, that is when the spinner is dismissed.

To make it easy/possible to use this with Relay, we need to expose async methods for refetching data as an alternative to the existing methods. This means adding a new async method on the wrapped value for `@Query` to refetch, and updating the existing refetch method for pagination and refetch fragments to be async.

For queries, extra care needed to be taken to avoid clearing out existing data when refetching. Progress is already shown by the pull to refresh UI, so it doesn't make sense to replace the app UI with a loading state. Instead the UI keeps showing the existing data until the refetch request is complete.

Test Plan: Use RelayStarWars as a test for this by changing its existing refresh button to use `refreshable`.

Reviewers: matt

Reviewed By: matt

Differential Revision: https://code.home.mattmoriarity.com/D14